test: add `test_list_user_workspaces`
